Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/invalid-uncompiled-raw-handler-keeps-position.window-expected.txt

CONSOLE MESSAGE: SyntaxError: Parser error

PASS Compiling invalid uncompiled raw handlers should keep the position in event listener list
